logic for declaring and retreving variable changed. scope number added with each variable.
for variable declaration: checking whether the variable is present in
the current or parent scope without checking the corresponding value. if
it returns false report error.

for retreving: if the variable value in the current scope is undefined
then check in the parent scope. if the result is false report error.

a = 10
{
b = a + 2
}
for this variable data will be:
a(0): 10
b(1): 12